Scripts for menu
Things are a bit spazzy still. Had no idea this would be such a pain in the hole.
Nevertheless, the object will slerp back to original pos if in any other position. Also adds a force to the spin based on distance of swipe with reagrds to time.
Currently only adds a a force to spin to the left. Will fix that later